moved dowloadcsv to component
